Major update for the V3 PCB
Major update for the V3 PCB: the cut-down is now based on the eRIC4 or eRIC9 radio controller and can be triggered by radio; a u-blox CAM-M8Q provides altitude information; the Shark and Servo are now mounted directly on the PCB instead of carbon fibre. The V1 files are in the Archive folder.